+++ packages/taro-components/types/Progress.d.ts
@@ -2,60 +2,60 @@ import { ComponentType } from 'react'
 import { StandardProps } from './common'
 
 interface ProgressProps extends StandardProps {
-  
+
   /**
    * 百分比0~100
    *
    * 默认值：`无`
    */
   percent: number,
-  
+
   /**
-   * 在进度条右侧显示百分比	
+   * 在进度条右侧显示百分比
    *
    * 默认值：`false`
    */
-  showInfo: boolean,
-  
+  showInfo?: boolean,
+
   /**
-   * 进度条线的宽度，单位px	
+   * 进度条线的宽度，单位px
    *
    * 默认值：`6`
    */
-  strokeWidth: number,
-  
+  strokeWidth?: number,
+
   /**
-   * 进度条颜色 （请使用 activeColor）	
+   * 进度条颜色 （请使用 activeColor）
    *
-   * 默认值：`09BB07`
+   * 默认值：`#09BB07`
    */
-  color: string,
-  
+  color?: string,
+
   /**
-   * 已选择的进度条的颜色	
+   * 已选择的进度条的颜色
    */
-  activeColor: string,
-  
+  activeColor?: string,
+
   /**
-   * 未选择的进度条的颜色	
+   * 未选择的进度条的颜色
    */
-  backgroundColor: string,
-  
+  backgroundColor?: string,
+
   /**
-   * 进度条从左往右的动画	
+   * 进度条从左往右的动画
    *
    * 默认值：`false`
    */
-  active: boolean,
-  
+  active?: boolean,
+
   /**
    * backwards: 动画从头播
-   * 
-   * forwards：动画从上次结束点接着播	
+   *
+   * forwards：动画从上次结束点接着播
    *
    * 默认值：`backwards`
    */
-  activeMode: 'backwards' | 'forwards'
+  activeMode?: 'backwards' | 'forwards'
 }
 
 declare const Progress: ComponentType<ProgressProps>